+++ src/utils.ts
@@ -1,0 +1,1 @@
+// Put any shared utility functions here if needed
